WIP: Node.show()
